Mock db connection
